Add aroles to endpoints

+++ FacebookClone/FacebookClone.Presentation/EndpointDefinitions/AlbumEndpointDefinition.cs
@@ -14,9 +14,9 @@ namespace FacebookClone.Presentation.EndpointDefinitions
     {
         public static void DefineEndpoints(WebApplication app)
         {
-            app.MapGet("/albums", [Authorize(Policy = "RequireId")] (IAlbumService albumService) => albumService.GetAll());
+            app.MapGet("/albums", [Authorize(Roles = "Admin", Policy = "RequireId")] (IAlbumService albumService) => albumService.GetAll());
 
-            app.MapPost("/albums", [Authorize(Policy = "RequireId")] (HttpRequest request, IAlbumService albumService, IWebHostEnvironment environment) =>
+            app.MapPost("/albums", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (HttpRequest request, IAlbumService albumService, IWebHostEnvironment environment) =>
             {
                 if (request.Form.Files.Count == 0)
                 {
@@ -53,15 +53,15 @@ namespace FacebookClone.Presentation.EndpointDefinitions
                 return result;
             });
 
-            app.MapGet("/albums/{id}", [Authorize(Policy = "RequireId")] (IAlbumService albumService, int id) => albumService.GetById(id));
+            app.MapGet("/albums/{id}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IAlbumService albumService, int id) => albumService.GetById(id));
 
-            app.MapGet("/albums/search/{userId}", [Authorize(Policy = "RequireId")] (IAlbumService albumService, int userId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllAlbumWithImagesByUserId(userId, pageSize, pageNumber));
+            app.MapGet("/albums/search/{userId}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IAlbumService albumService, int userId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllAlbumWithImagesByUserId(userId, pageSize, pageNumber));
 
-            app.MapDelete("/albums/{id}", [Authorize(Policy = "RequireId")] (IAlbumService albumService, int id) => albumService.Delete(id));
+            app.MapDelete("/albums/{id}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IAlbumService albumService, int id) => albumService.Delete(id));
 
-            app.MapPut("/albums", [Authorize(Policy = "RequireId")] (AlbumDTO album, IAlbumService albumService) => albumService.Update(album));
+            app.MapPut("/albums", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (AlbumDTO album, IAlbumService albumService) => albumService.Update(album));
 
-            app.MapGet("/feed", [Authorize(Policy = "RequireId")] (HttpContext context, IAlbumService albumService, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllFriendsWithAlbumsWithImages(Convert.ToInt32(context.User.Claims.SingleOrDefault(e => e.Type == "id").Value), pageSize, pageNumber));
+            app.MapGet("/feed", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (HttpContext context, IAlbumService albumService, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllFriendsWithAlbumsWithImages(Convert.ToInt32(context.User.Claims.SingleOrDefault(e => e.Type == "id").Value), pageSize, pageNumber));
 
         }
     }
