+++ FacebookClone/FacebookClone.Presentation/EndpointDefinitions/UserEndpointDefinition.cs
@@ -13,24 +13,24 @@ namespace FacebookClone.Presentation.EndpointDefinitions
     {
         public static void DefineEndpoints(WebApplication app)
         {
-            app.MapGet("/users", [Authorize(Policy = "RequireId")] (IUserService userService) => userService.GetAll());
+            app.MapGet("/users", [Authorize(Roles = "Admin", Policy = "RequireId")] (IUserService userService) => userService.GetAll());
 
             app.MapPost("/register", (RegisterDTO userRegister, IUserService userService) => userService.Add(userRegister));
 
             app.MapPost("/login", (LoginDTO userLogin, IJwtTokenService jwtTokenService) => jwtTokenService.GenerateJwt(userLogin));
 
-            app.MapGet("/home", [Authorize(Policy = "RequireId")] (IUserService userService, HttpContext context) =>
+            app.MapGet("/home", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IUserService userService, HttpContext context) =>
             {
                 return Results.Ok(userService.GetById(Convert.ToInt32(context.User.Claims.SingleOrDefault(e => e.Type == "id").Value)));
             });
 
-            app.MapGet("/users/{id}", [Authorize(Policy = "RequireId")] (IUserService userService, int id) => userService.GetById(id));
+            app.MapGet("/users/{id}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IUserService userService, int id) => userService.GetById(id));
 
-            app.MapGet("/users/search/{username}", [Authorize(Policy = "RequireId")] (IUserService userService, string username) => userService.SearchByUsername(username));
+            app.MapGet("/users/search/{username}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IUserService userService, string username) => userService.SearchByUsername(username));
 
-            app.MapDelete("/users/{id}", [Authorize(Policy = "RequireId")] (IUserService userService, int id) => userService.Delete(id));
+            app.MapDelete("/users/{id}", [Authorize(Roles = "Admin", Policy = "RequireId")] (IUserService userService, int id) => userService.Delete(id));
 
-            app.MapPut("/updateProfileImage", [Authorize(Policy = "RequireId")] (HttpRequest request, HttpContext context, IUserService userService, IWebHostEnvironment environment) =>
+            app.MapPut("/updateProfileImage", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (HttpRequest request, HttpContext context, IUserService userService, IWebHostEnvironment environment) =>
             {
                 int userId = Convert.ToInt32(context.User.Claims.SingleOrDefault(e => e.Type == "id").Value);
 
@@ -48,7 +48,7 @@ namespace FacebookClone.Presentation.EndpointDefinitions
                 return userService.UpdateProfileImage(userId, imageUrl,environment.WebRootPath);
             });
 
-            app.MapPut("/updateCoverImage", [Authorize(Policy = "RequireId")] (HttpRequest request, HttpContext context, IUserService userService, IWebHostEnvironment environment) =>
+            app.MapPut("/updateCoverImage", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (HttpRequest request, HttpContext context, IUserService userService, IWebHostEnvironment environment) =>
             {
                 int userId = Convert.ToInt32(context.User.Claims.SingleOrDefault(e => e.Type == "id").Value);
 
@@ -66,7 +66,7 @@ namespace FacebookClone.Presentation.EndpointDefinitions
                 return userService.UpdateCoverImage(userId, imageUrl, environment.WebRootPath);
             });
 
-            app.MapPut("/banUser/{id}", [Authorize(Policy = "RequireId")] (IUserService userService, int id) => userService.BanUserById(id));
+            app.MapPut("/banUser/{id}", [Authorize(Roles = "Admin", Policy = "RequireId")] (IUserService userService, int id) => userService.BanUserById(id));
 
             app.MapGet("/confirmMail/{emailHash}", (HttpResponse response, IEmailConfirmService emailConfirmService, string emailHash) =>
             {
